works with dask delayed

+++ nexoclom/modelcode/LOSResultFitted.py
@@ -75,17 +75,20 @@ class LOSResultFitted(LOSResult):
             # Check to see if there is already a result for this
             search_result = self.fitted_iteration_search(ufit_id)
             
-            if overwrite:
+            if overwrite and (search_result is not None):
                 metadata_obj = sqla.MetaData()
                 table = sqla.Table("uvvsmodels", metadata_obj, autoload_with=engine)
                 del_stmt = sqla.delete(table).where(
-                    table.columns.idnum == search_result[0])
+                    table.columns.idnum == int(search_result[0]))
 
                 with engine.connect() as con:
                     con.execute(del_stmt)
                     con.commit()
+                    
                 if os.path.exists(search_result[2]):
                     os.remove(search_result[2])
+                else:
+                    pass
                 search_result = None
             else:
                 pass
@@ -107,6 +110,9 @@ class LOSResultFitted(LOSResult):
                 packets = output.X.copy()
                 packets0 = output.X0.copy()
                 
+                # radiance = fitted radiance
+                # weighting = weighting factor for each X0
+                # included = number of times trajectory included in lines of sight
                 radiance = pd.Series(np.zeros(data.shape[0]), index=data.index)
                 weighting = pd.Series(np.zeros(packets0.shape[0]),
                                       index=packets0.index)
@@ -115,18 +121,31 @@ class LOSResultFitted(LOSResult):
                 
                 ratio = data.radiance / unfit_model_result.radiance
                 ratio.fillna(0, inplace=True)
-                
+
                 for spnum, spectrum in data.iterrows():
                     used = list(iteration_unfit.used_packets.loc[spnum])
                     cts = packets.loc[used, 'Index'].value_counts()
                     weighting.loc[cts.index] += cts.values * ratio[spnum]
                     included.loc[cts.index] += cts.values
-                    
+
                 used = included > 0
                 weighting[used] = weighting[used] / included[used]
                 weighting /= weighting[used].mean()
                 assert np.all(np.isfinite(weighting))
                 
+                # def find_where_used(ind):
+                #     which = iteration_unfit.used_packets0.apply(lambda x:ind in x)
+                #     return list(iteration_unfit.used_packets0.index[which])
+
+                # used = pd.Series(output.X0.index).apply(find_where_used)
+                # weight = used.apply(lambda x:ratio[x].values)
+                # weighting = weight.apply(lambda x: np.mean(x) if len(x) > 0 else 0)
+
+                # from inspect import currentframe, getframeinfo
+                # frameinfo = getframeinfo(currentframe())
+                # print(frameinfo.filename, frameinfo.lineno)
+                # from IPython import embed; embed()
+
                 multiplier = weighting.loc[output.X['Index']].values
                 output.X.loc[:, 'frac'] = output.X.loc[:, 'frac'] * multiplier
                 output.X0.loc[:, 'frac'] = output.X0.loc[:, 'frac'] * weighting
@@ -154,8 +173,8 @@ class LOSResultFitted(LOSResult):
                         pass
 
                 # Save the fitted output
-                # output.inputs = self.inputs
-                # output.save()
+                output.inputs = self.inputs
+                output.save()
                 
                 iteration = {'radiance': radiance.values,
                              'npackets': output.X0.frac.sum(),
@@ -180,17 +199,19 @@ class LOSResultFitted(LOSResult):
                 iteration_result.modelfile = search_result[2]
                 fitted_iteration_results.append(iteration_result)
             self.modelfiles = {}
-            
+
+        self.outputfiles = []
         for iteration_result in fitted_iteration_results:
             self.radiance += iteration_result.radiance
             self.totalsource += iteration_result.totalsource
-            self.modelfiles[iteration_result.outputfile] = iteration_result.modelfile
+            self.modelfiles[iteration_result.unfit_outputfile] = iteration_result.modelfile
+            self.outputfiles.append(iteration_result.outputfile)
         
         model_rate = self.totalsource/self.inputs.options.endtime.value
         self.atoms_per_packet = 1e23 / model_rate
         self.radiance *= self.atoms_per_packet/1e3*u.kR
         self.determine_source_rate(scdata)
         self.atoms_per_packet *= self.sourcerate.unit
-        self.outputfiles = list(self.modelfiles.keys())
+        self.unfit_outputfiles = list(self.modelfiles.keys())
 
         print(self.totalsource, self.atoms_per_packet)
